Move the model into nemotron_vl.py
